% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_MLE.R
\docType{data}
\name{DM_MLE}
\alias{DM_MLE}
\title{Baseline parameter for true compositional abundance}
\format{
A compositional vector with dimension P=1198 without annotation.
}
\usage{
DM_MLE
}
\description{
Baseline parameter for true compositional abundance
}
\keyword{datasets}
