% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{A real data example}
\format{
Dataframes for clinical variables and OTUs' relative abundance table.
}
\usage{
example_data
}
\description{
A real data example
}
\keyword{datasets}
