% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMR_Tab.R
\name{JMR_Tab}
\alias{JMR_Tab}
\title{Joint model with Matching and Regularization (JMR)}
\usage{
JMR_Tab(
  otu_tab,
  long_design,
  logistic_design,
  outcome,
  long_idset,
  logistic_idset,
  rand.var,
  tune = seq(0.05, 0.15, 0.05),
  cov.taxa = T
)
}
\arguments{
\item{otu_tab}{A table of relative abundance with rows as OTUs and columns as samples.}

\item{long_design}{A numeric vector for taxa filters, specified in the order of c(minimum relative abundance, minimum prevalence )}

\item{logistic_design}{A character vector for set and subject identifiers in DataPrep$meta_data, in the order of c(set ID, subject ID)}

\item{outcome}{A character for disease outcome variable in DataPrep$meta_data.}

\item{long_idset}{Name of the covariate representing age (or time point) for each sample.}

\item{logistic_idset}{Names of the covariates in disease sub-model.}

\item{rand.var}{Names of the covariates used in OTU sub-models.}

\item{tune}{The type of trajectory analysis: intercept or slope.}

\item{cov.taxa}{Whether to adjust for covariate taxa.}
}
\value{
\item{$test.result}{The result of joint test on relative abundance and presence.}
\item{$rho}{Tuning parameter value.}
}
\description{
This function test trajectory (either intercept or slope) association with disease outcome in matched sets.
}
