% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointMatch.r
\name{JointMatch}
\alias{JointMatch}
\title{Joint nested random effect model with Matched sets indicator}
\usage{
JointMatch(
  DataPrep,
  filters,
  identifiers,
  disease_status,
  sample_age = NULL,
  disease_cov,
  OTU_cov,
  trajectory_type,
  shrinkage = 0.2,
  matching_type = NULL,
  matching_factors = NULL,
  trace = TRUE
)
}
\arguments{
\item{DataPrep}{A list of object returned by DataPrep function or prepared by users. This object must contain $relabun: a relative abundance
matrix or data frame with rows as taxa and columns as samples; $meta_data: a matrix or data frame of subject ID, set ID, sample ID, disease outcome, time-variant and time-invariant covariates for all submodels,
with columns as variables, and rows as samples in the order of $relabun's samples (columns).}

\item{filters}{A numeric vector for taxa filters, specified in the order of c(minimum relative abundance, minimum prevalence )}

\item{identifiers}{A character vector for set and subject identifiers in DataPrep$meta_data, in the order of c(set ID, subject ID)}

\item{disease_status}{A character for disease outcome variable in DataPrep$meta_data.}

\item{sample_age}{Name of the covariate representing age (or time point) for each sample.}

\item{disease_cov}{Names of the covariates in disease sub-model.}

\item{OTU_cov}{Names of the covariates used in OTU sub-models.}

\item{trajectory_type}{The type of trajectory analysis: intercept or slope.}

\item{shrinkage}{Regularization parameter defaults to 0.2.}

\item{matching_type}{If participants are matched by a nested case-control design, set matching_type='ncc'.Otherwise, matching_type is NULL as default.}

\item{matching_factors}{If matching_type='ncc', matching_factors must be provided.}

\item{trace}{If trace=FALSE, JointMatch does not print out parameter estimate per taxon. Default to TRUE.}
}
\value{
\item{$OutputPrt}{A list of analysis result for each submodel per taxon. Rows annotated as 'nzAbundance:' is the result for the submodel of non-zero abundance;
rows annotated as 'Presence:' is the result for the submodel of presence; rows annotated as 'Disease:' is the result for the submodel of disease risk.}
\item{$IndividualTest}{The test of taxon-disease association in non-zero abundance and presence, individually}
\item{$JointTest}{Jointly test if a taxon is associated with disease outcome in either non-zero abundance or presence}
}
\description{
This function performs trajectory analysis (either intercept or slope) with disease outcome in matched sets.
}
