% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointSingle.R
\name{JointSingle}
\alias{JointSingle}
\title{Joint Single random effect model}
\usage{
JointSingle(
  DataPrep,
  filters,
  identifiers,
  disease_status,
  sample_age = NULL,
  disease_cov,
  OTU_cov,
  trajectory_type,
  shrinkage = 0.2,
  matching_type = NULL,
  matching_factors = NULL,
  trace = TRUE
)
}
\arguments{
\item{DataPrep}{An object returned by DataPrep function.}

\item{filters}{A numeric vector for taxa filters, specified in the order of c(minimum relative abundance, minimum prevalence )}

\item{disease_status}{A character for disease outcome variable in DataPrep$meta_data.}

\item{sample_age}{Name of the covariate representing age (or time point) for each sample.}

\item{disease_cov}{Names of the covariates in disease sub-model.}

\item{OTU_cov}{Names of the covariates used in OTU sub-models.}

\item{trajectory_type}{The type of trajectory analysis: intercept or slope.}

\item{shrinkage}{Regularization parameter defaults to 0.2.}

\item{trace}{If trace=FALSE, JointMatch does not print out parameter estimate per taxon. Default to TRUE.}

\item{identifier}{A character vector for identifiers in DataPrep$meta_data, in the order of c(set indicator, subject ID).}
}
\description{
This function performs trajectory analysis (either intercept or slope) with disease outcome, not using matched set indicator.
}
