% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatSim.R
\name{StatSim}
\alias{StatSim}
\title{Simulation for disease status, risk factor(s) and time points in matched sets}
\usage{
StatSim(n, omegas = c(1.5, 1), tps = 5)
}
\arguments{
\item{n}{Number of subjects}

\item{omegas}{The values of omegas for subject-level and set-level effect on disease}

\item{tps}{Number of time points.}
}
\description{
Simulation for disease status, risk factor(s) and time points in matched sets
}
