% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxaSim.R
\name{TaxaSim}
\alias{TaxaSim}
\title{Simulation for a temporal high-dimensional OTU counts table based on StatSim output}
\usage{
TaxaSim(
  base_par,
  StatSim,
  shift_subject = 0,
  library.size = 1e+05,
  min.lib = 10000,
  trace = TRUE
)
}
\arguments{
\item{base_par}{A high-dimensional compositional vector as Dirichlet-Multinomial baseline parameter.}

\item{StatSim}{A data matrix of disease status, identifiers, and the other covariates, generated by StatSim().}

\item{shift_subject}{Shift size between two subjects with value between 0 and 1.}

\item{library.size}{The mean level of metagenomic sequencing library size.}

\item{min.lib}{The minimum threshold for sample-wise library size.}

\item{trace}{Prints out the simulated true relative abundance of the first 10 taxa. Defaults to TRUE.}
}
\description{
Simulation for a temporal high-dimensional OTU counts table based on StatSim output
}
