% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.r
\name{DataPrep}
\alias{DataPrep}
\title{Data preprocessing}
\usage{
DataPrep(counts, level, covdata, sample_id, composition, cleananno)
}
\arguments{
\item{counts}{A data frame of OTU-level raw counts with rows as taxa and columns as samples.}

\item{level}{One or multiple levels in 'kingdom','phylum','class','order','family','genus','species'. Default to 'species'.}

\item{covdata}{A data frame of all the covariates to be included in trajectory analysis, either longitudinal or non-longitudinal.}

\item{sample_id}{Specify the variable for microbiome sample ID in covdata, which should be identical to the column names of counts.}

\item{composition}{If composition = TRUE, then composition. Otherwise, return raw counts in output. Default to TRUE and return counts only.}

\item{cleananno}{If cleananno=TRUE, then it cleans out special characters from OTU names.}
}
\description{
This function transforms the raw counts OTU table into compositional abundance table based on the specified taxonomy level(s)
}
