% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointMatch.r
\name{JointMatch}
\alias{JointMatch}
\title{Joint nested random effect model with Matched sets indicator}
\usage{
JointMatch(
  DataPrep,
  filters,
  identifiers,
  disease_status,
  sample_age = NULL,
  disease_cov,
  OTU_cov,
  trajectory_type,
  shrinkage = 0.2,
  matching_type = NULL,
  matching_factors = NULL,
  trace = TRUE
)
}
\arguments{
\item{DataPrep}{An object returned by DataPrep function}

\item{filters}{A numeric vector for taxa filters, specified in the order of c(minimum relative abundance, minimum prevalence )}

\item{identifiers}{A character vector for identifiers in DataPrep$meta_data, in the order of c(set indicator, subject ID)}

\item{disease_status}{A character for disease outcome variable in DataPrep$meta_data.}

\item{sample_age}{Name of the covariate representing age (or time point) for each sample.}

\item{disease_cov}{Names of the covariates in disease sub-model.}

\item{OTU_cov}{Names of the covariates used in OTU sub-models.}

\item{trajectory_type}{The type of trajectory analysis: intercept or slope.}

\item{shrinkage}{Regularization parameter defaults to 0.2.}

\item{matching_type}{If participants are matched by a nested case-control design, set matching_type='ncc'.Otherwise, matching_type is NULL as default.}

\item{matching_factors}{If matching_type='ncc', matching_factors must be provided.}

\item{trace}{If trace=FALSE, JointMatch does not print out parameter estimate per taxon. Default to TRUE.}
}
\description{
This function performs trajectory analysis (either intercept or slope) with disease outcome in matched sets.
}
